- Added more configurable classes
