+++ src/Utils/GeneratorForeignKey.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace InfyOm\Generator\Utils;
+
+
+class GeneratorForeignKey
+{
+    /** @var string */
+    public $name;
+    public $localField;
+    public $foreignField;
+    public $foreignTable;
+    public $onUpdate;
+    public $onDelete;
+}
